+++ examples/26_map.cc
@@ -41,7 +41,7 @@
 
 static constexpr int GridSize = 60;
 static constexpr float CellSize = 10.0f;
-static constexpr int Size = GridSize * CellSize;
+static constexpr int Size = static_cast<int>(GridSize * CellSize);
 
 static const char *ExampleMap[GridSize] = {
   "############################################################",
